+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "files.associations": {
+        "gaslight-sort.h": "c"
+    }
+}
