+++ C++/Gaslight-Sort.cpp
@@ -1,0 +1,13 @@
+#include <iostream>
+#include "Gaslight-Sort.h"
+
+int main()
+{
+    
+    return 0;
+}
+
+bool isSorted(int *array)
+{
+	return true;
+}
