+++ C++/Gaslight-sort.h
@@ -1,0 +1,6 @@
+#ifndef GASLIGHT_H
+#define GASLIGHT_H
+
+bool isSorted(int *array);
+
+#endif
